Cleaning and moving props part 9
